Bluetooth: Test: Split up test "uuid"
Create a separate file and test suite for each tested function.
